Convert dat.gz file to Apache Arrow Datasets using partitioned folder structures

Some large files in the SRE are too large to fit into memory. dat_to_datasets reads files into memory in smaller chunks (controlled by the chunk_size argument) and converts them into Arrow Datasets. All ... argument are passed to dipr::read_dat which is where column types can be specified.

dat_to_datasets(
  data_path,
  data_dict,
  chunk_size = 1e+06,
  path,
  partitioning,
  ...
)

Arguments

data_path

A path or a vector of paths to a .dat.gz file. If supplying a vector of paths, they must share a common data dictionary.

data_dict

A data.frame with start, stop and name columns

chunk_size

The number of rows to include in each chunk. The value of this parameter you choose will depend on both the number of rows in the data you are trying to process and the RAM available. You can check the RAM available using memory.size(max = TRUE). The default for this values is currently 10 million.

path

string path, URI, or SubTreeFileSystem referencing a directory to write to (directory will be created if it does not exist)

partitioning

Partitioning or a character vector of columns to use as partition keys (to be written as path segments). Default is to use the current group_by() columns.

...

Arguments passed on to readr::read_fwf

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector).

Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded and decompressed.

Literal data is most useful for examples and tests. To be recognised as literal data, the input must be either wrapped with I(), be a string containing at least one new line, or be a vector containing at least one string with a new line.

Using a value of clipboard() will read from the system clipboard.

col_positions

Column positions, as created by fwf_empty(), fwf_widths() or fwf_positions(). To read in only selected fields, use fwf_positions(). If the width of the last column is variable (a ragged fwf file), supply the last end position as NA.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from guess_max rows on the input interspersed throughout the file. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

col_select

<tidy-select> Columns to include in the results, either by name or by numeric index. Use c() or list() to select with more than one expression and ?tidyselect::language for full details on the selection language.

id

The name of a column in which to store the file path. This is useful when reading multiple input files and there is data in the file paths, such as the data collection date. If NULL (the default) no extra column is created.

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

na

Character vector of strings to interpret as missing values. Set this option to character() to indicate no missing values.

comment

A string used to identify comments. Any text after the comment characters will be silently ignored.

trim_ws

Should leading and trailing whitespace (ASCII spaces and tabs) be trimmed from each field before parsing it?

skip

Number of lines to skip before reading data.

n_max

Maximum number of lines to read.

guess_max

Maximum number of lines to use for guessing column types.

progress

Display a progress bar? By default it will only display in an interactive session and not while knitting a document. The automatic progress bar can be disabled by setting option readr.show_progress to FALSE.

name_repair

Treatment of problematic column names:

"minimal": No name repair or checks, beyond basic existence of names

"unique": Make sure names are unique and not empty

"check_unique": (default value), no name repair, but check they are unique

"universal": Make the names unique and syntactic

a function: apply custom name repair (e.g., .name_repair = make.names for names in the style of base R)

A purrr-style anonymous function, see rlang::as_function()

This argument is passed on as repair to vctrs::vec_as_names(). See there for more details on these terms and the strategies used to enforce them.

num_threads

The number of processing threads to use for initial parsing and lazy reading of data. If your data contains newlines within fields the parser should automatically detect this and fall back to using one thread only. However if you know your file has newlines within quoted fields it is safest to set num_threads = 1 explicitly.

show_col_types

If FALSE, do not show the guessed column types. If TRUE always show the column types, even if they are supplied. If NULL (the default) only show the column types if they are not explicitly supplied by the col_types argument.

lazy

Read values lazily? By default the file is initially only indexed and the values are read lazily when accessed. Lazy reading is useful interactively, particularly if you are only interested in a subset of the full dataset. Note, if you later write to the same file you read from you need to set lazy = FALSE. On Windows the file will be locked and on other systems the memory map will become invalid.

skip_empty_rows

Should blank rows be ignored altogether? i.e. If this option is TRUE then blank rows will not be represented at all. If it is FALSE then they will be represented by NA values in all the columns.

Examples

data_dict_path <- dipr_example("starwars-dict.txt")
dict <- read.table(data_dict_path)
dat_path <- dipr_example("starwars-fwf.dat.gz")

## Create a partitioned datasets in the "bar" folder
dat_to_datasets(
    data_path = dat_path,
    data_dict = dict,
    path = "starwars_arrow",
    partitioning = "species",
    chunk_size = 2)
#> Rows: 2 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> 
#> chr (2): name, species
#> dbl (2): height, mass
#> lgl (1): has_hair
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 2 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> 
#> chr (2): name, species
#> dbl (2): height, mass
#> lgl (1): has_hair
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 2 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> 
#> chr (2): name, species
#> dbl (2): height, mass
#> lgl (1): has_hair
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 2 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> 
#> chr (2): name, species
#> dbl (2): height, mass
#> lgl (1): has_hair
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 2 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> 
#> chr (2): name, species
#> dbl (2): height, mass
#> lgl (1): has_hair
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 2 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> 
#> chr (2): name, species
#> dbl (2): height, mass
#> lgl (1): has_hair
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 2 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> 
#> chr (2): name, species
#> dbl (2): height, mass
#> lgl (1): has_hair
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 1 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> 
#> chr (2): name, species
#> dbl (2): height, mass
#> lgl (1): has_hair
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.