An opinionated template for package-like targets project

This function will create a DESCRIPTION file, a _targets.R template and an R directory to store function files.

dipr_create_targets_project(...)

Arguments

...

Arguments passed on to usethis::use_description

fields

A named list of fields to add to DESCRIPTION, potentially overriding default values. See use_description() for how you can set personalized defaults using package options

check_name

Whether to check if the name is valid for CRAN and throw an error if not

roxygen

If TRUE, sets RoxygenNote to current roxygen2 version