Assign a category to a vector of icd codes

Vectorized function that assigns categories to icd codes. Thin wrappers are provided for convenience for ICD-9 and ICD-10 codes respectively. Other than the icd_version argument, each wrapper accepts the same arguments as the core function.

dipr_icd_categories(
  icd_codes = NULL,
  icd_version = 10,
  include_code_ranges = FALSE,
  mortality = FALSE,
  print_defs_short = FALSE,
  print_defs_long = FALSE,
  export_defs = FALSE
)

dipr_icd9_categories(...)

dipr_icd10_categories(...)

Arguments

icd_codes

vector of icd codes

icd_version

version of the icd code system

include_code_ranges

include codes used to assign category in returned vector.

mortality

default is FALSE. Set to TRUE if working with mortality data

print_defs_short

print a short definition

print_defs_long

print a long definition

export_defs

write out a csv. Defaults to FALSE.

...

arguments passed to dipr_icd_categories

Details

ICD-10 categories were defined according to the official 2019 WHO ICD-10 website: https://icd.who.int/browse10/2019/en

ICD-9 categories were defined according to the BC Ministry of Health's ICD-9 Diagnostic Code Descriptions website: https://www2.gov.bc.ca/gov/content/health/practitioner-professional-resources/msp/physicians/diagnostic-code-descriptions-icd-9

Functions

dipr_icd9_categories: Thin wrapper around dipr_icd_categories for only ICD-9 codes

dipr_icd10_categories: Thin wrapper around dipr_icd_categories for only ICD-10 codes

Author

Craig Hutton (craig.hutton@gov.bc.ca)

Examples

## ICD-9
dipr_icd9_categories("C19")
#> [1] "SARS/COVID-19"

## ICD-10
dipr_icd10_categories("U07.1")
#> [1] "SARS/COVID-19"