Get the dat path of a particular dataset for a given core table name

Return the paths of all data within a version that match a table name

get_core_dat_path(
  core_version,
  core_table_name = NULL,
  core_table_pattern = NULL,
  pattern = "\\.dat.gz$",
  base_path = "R:/DATA/core-snapshot"
)

Arguments

core_version

Core version of tables typically formatted as a YYYYMMDD string

core_table_name

an exact string of the data set name. Should match the folder inside the core-snapshot path. Leave as NULL if using core_table_pattern to match several folders.

core_table_pattern

a string (can be a regex) to match a table name - can be a subdataset of core_table_name or a pattern to match a dataset that may be split over multiple dataset folders in the core-snapshot path.

pattern

Defaults to dat.gz but can be extended for more extended using regex for more flexibility

base_path

Defaults to "R:/DATA/core-snapshot"