Process msp data and return unique visits

This method takes the raw msp input object (using read_dat) and returns a condensed data.table that has one row per unique visit to a doctor or hospital. This function requires data to be in msp.c format with those column heading.

msp_unique(msp_c, primary_icd = FALSE)

Arguments

msp_c

A data.table object of msc.c status.

primary_icd

Default FALSE. Set to TRUE is memory issues are occuring. This feature returns only the primary ICD 9 code

Author

Jenny Sutherland, Sam Albers