Prepare repo for import into the SRE through OCWA

Removes files listed in _ocwaignore and processes README.md so that it is not using images or URLs that can't be accessed in the SRE.

ocwa_branch_export(branch = "ocwa-import", ask = TRUE)

Arguments

branch

branch name to prepare for OCWA. Default "ocwa-import"

ask

should the function stop and ask for verification to proceed? Default TRUE

Value

hash of commit

Details

This function:

Creates a new branch

Removes files that can't be imported - these are listed in the _ocwaignore file in the root of the repo

Cleans README.md to comment out references to images and links that won't be available in the SRE.

Commits the changes from 2 and 3 to the new branch and pushing that to GitHub. This branch can then be used as the basis for an import into OCWA.